[Decomp][Chat] Cohesion split: move list-display helpers to ListCommands.cpp
